Binomial Distribution Experiment (C++)
This C++ code file is an attempt to create the binomial distribution formula through 3 functions. 2 functions compute the factorial and the binomial coefficient which are essential to calculate the binomial distribution which is calculated by the third and final function. In the main function, I inserted a line of code that outputs the binomial distribution using three inputs that can be changed by the user. The binomial distribution is used to calculate probabilities, and is one method that is integrated in finance to calculate future stock prices which is one of the reasons I created this file.